Use SyncedList, add UI for add item and change item

+++ ShoppingListAndroid/src/main/java/de/evylon/shoppinglist/android/gui/utils/FloatUtils.kt
@@ -2,10 +2,10 @@ package de.evylon.shoppinglist.android.gui.utils
 
 import kotlin.math.roundToInt
 
-fun Float.prettyFormat(): String = when {
+fun Double.prettyFormat(): String = when {
     this.isDecimal -> this.toInt().toString()
     else -> "%.2f".format(this)
 }
 
-val Float.isDecimal: Boolean
-    get() = this.roundToInt().toFloat() == this
+val Double.isDecimal: Boolean
+    get() = this.roundToInt().toDouble() == this
